requirements.txt
langchain-core
langchain
langsmith
pdfminer.six
modal-client
fastapi-poe
google-cloud-storage
langchain-openai
openai
langchain-mistralai
langchain-google-vertexai
google-cloud-bigquery
google-cloud-translate
google-cloud-vision
firebase-admin